refactor: Make package for basic class
