+++ src/main/java/nextstep/ladder/Player.java
@@ -1,5 +1,8 @@
 package nextstep.ladder;
 
+import nextstep.ladder.model.Name;
+import nextstep.ladder.model.Position;
+
 public class Player {
   private Name name;
   private Position position;
