+++ src/test/java/nextstep/ladder/model/NameTest.java
@@ -1,7 +1,8 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import nextstep.ladder.model.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
